bugfix in image::addFunctionStub, compile-time fixes in pdwinnt.C and in libcommon   - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - removed unnecessary typecast that was causing a compile-time error

+++ common/src/serialize-xml.C
@@ -195,7 +195,7 @@ void SerDesXML::translate(int &param, const char *tag)
 
 void SerDesXML::translate(unsigned int &param, const char *tag)
 {   
-  translate( (int) param, tag);
+  translate( param, tag);
 }
 
 #if 0
